fix circular dependency between count_query_runner.ts and lite-api/aggregate.ts
