+++ packages/firestore/src/api/aggregate.ts
@@ -17,21 +17,23 @@
 
 import { Query } from '../api';
 import { firestoreClientRunCountQuery } from '../core/firestore_client';
-import { AggregateField, AggregateQuerySnapshot } from '../lite-api/aggregate';
+import { AggregateField, AggregateQuerySnapshot } from '../lite-api/aggregate_types';
 import { cast } from '../util/input_validation';
 
 import { ensureFirestoreConfigured, Firestore } from './database';
 import { ExpUserDataWriter } from './reference_impl';
 
 export {
+  aggregateQuerySnapshotEqual
+} from '../lite-api/aggregate';
+export {
   AggregateField,
   AggregateFieldType,
   AggregateSpec,
   AggregateSpecData,
   AggregateQuerySnapshot,
-  count,
-  aggregateQuerySnapshotEqual
-} from '../lite-api/aggregate';
+  count
+} from '../lite-api/aggregate_types';
 
 /**
  * Executes the query and returns the results as a `AggregateQuerySnapshot` from the
