+++ packages/firestore/src/lite-api/aggregate.ts
@@ -19,75 +19,11 @@ import { deepEqual } from '@firebase/util';
 import { CountQueryRunner } from '../core/count_query_runner';
 import { cast } from '../util/input_validation';
 
+import { AggregateField, AggregateQuerySnapshot, AggregateSpec } from './aggregate_types';
 import { getDatastore } from './components';
 import { Firestore } from './database';
 import { Query, queryEqual } from './reference';
 import { LiteUserDataWriter } from './reference_impl';
-
-/**
- * An `AggregateField`that captures input type T.
- */
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export class AggregateField<T> {
-  type = 'AggregateField';
-}
-
-/**
- * Creates and returns an aggregation field that counts the documents in the result set.
- * @returns An `AggregateField` object with number input type.
- */
-export function count(): AggregateField<number> {
-  return new AggregateField<number>();
-}
-
-/**
- * The union of all `AggregateField` types that are returned from the factory
- * functions.
- */
-export type AggregateFieldType = ReturnType<typeof count>;
-
-/**
- * A type whose values are all `AggregateField` objects.
- * This is used as an argument to the "getter" functions, and the snapshot will
- * map the same names to the corresponding values.
- */
-export interface AggregateSpec {
-  [field: string]: AggregateFieldType;
-}
-
-/**
- * A type whose keys are taken from an `AggregateSpec` type, and whose values
- * are the result of the aggregation performed by the corresponding
- * `AggregateField` from the input `AggregateSpec`.
- */
-export type AggregateSpecData<T extends AggregateSpec> = {
-  [P in keyof T]: T[P] extends AggregateField<infer U> ? U : never;
-};
-
-/**
- * An `AggregateQuerySnapshot` contains the results of running an aggregate query.
- */
-export class AggregateQuerySnapshot<T extends AggregateSpec> {
-  readonly type = 'AggregateQuerySnapshot';
-
-  /** @hideconstructor */
-  constructor(
-    readonly query: Query<unknown>,
-    private readonly _data: AggregateSpecData<T>
-  ) {}
-
-  /**
-   * The results of the requested aggregations. The keys of the returned object
-   * will be the same as those of the `AggregateSpec` object specified to the
-   * aggregation method, and the values will be the corresponding aggregation
-   * result.
-   *
-   * @returns The aggregation statistics result of running a query.
-   */
-  data(): AggregateSpecData<T> {
-    return this._data;
-  }
-}
 
 /**
  * Counts the number of documents in the result set of the given query, ignoring
